Security, CI and basic services

+++ qualitag/src/main/java/it/unisannio/studenti/qualitag/service/UserService.java
@@ -1,125 +1,38 @@
 package it.unisannio.studenti.qualitag.service;
 
-import it.unisannio.studenti.qualitag.dto.user.UserLoginDto;
 import it.unisannio.studenti.qualitag.dto.user.UserModifyDto;
-import it.unisannio.studenti.qualitag.dto.user.UserRegistrationDto;
 import it.unisannio.studenti.qualitag.mapper.UserMapper;
 import it.unisannio.studenti.qualitag.model.User;
 import it.unisannio.studenti.qualitag.repository.UserRepository;
+import it.unisannio.studenti.qualitag.security.model.CustomUserDetails;
+import it.unisannio.studenti.qualitag.security.service.AuthenticationService;
+import it.unisannio.studenti.qualitag.security.service.JwtService;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
 import jakarta.validation.ValidatorFactory;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
 import java.util.regex.Pattern;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 
 /**
  * The userService class is a service class that provides methods to manage the user entity.
  */
 @Service
+@RequiredArgsConstructor
 public class UserService {
-
-  private static final int LOG_ROUNDS = 12; // Adjust this value as needed
 
   private final UserRepository userRepository;
   private final UserMapper userMapper;
+  private final JwtService jwtService;
 
   private final ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
   private final Validator validator = factory.getValidator();
-
-  /**
-   * Constructs a new UserService.
-   *
-   * @param userRepository The user repository.
-   */
-  public UserService(UserRepository userRepository) {
-    this.userRepository = userRepository;
-    this.userMapper = new UserMapper(this);
-  }
-
-  /**
-   * Hashes a password using BCrypt.
-   *
-   * @param password The password to hash.
-   * @return The hashed password.
-   */
-  public String hashPassword(String password) {
-    return BCrypt.hashpw(password, BCrypt.gensalt(LOG_ROUNDS));
-  }
-
-  /**
-   * Checks if a password matches a hashed password.
-   *
-   * @param password The password to check.
-   * @param hashed   The hashed password to compare against.
-   * @return True if the password matches the hashed password, false otherwise.
-   */
-  public boolean checkPassword(String password, String hashed) {
-    return BCrypt.checkpw(password, hashed);
-  }
-
-  /**
-   * Checks if the email is valid.
-   *
-   * @param email The email to check.
-   * @return true if the email is valid, false otherwise.
-   */
-  public boolean isValidEmail(String email) {
-    String emailRegex = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$";
-    Pattern pattern = Pattern.compile(emailRegex);
-    return pattern.matcher(email).matches();
-  }
-
-  /**
-   * Checks if the password is valid. Minimum requirements:
-   * - At least one uppercase letter
-   * - At least one lowercase letter
-   * - At least one digit
-   * - At least one special character
-   * - Minimum length of 8 characters
-   *
-   * @param password The password to check.
-   * @return true if the password is valid, false otherwise.
-   */
-  public boolean isValidPassword(String password) {
-    String passwordRegex = "^(?=.*[A-Z])"
-        + "(?=.*[a-z])"
-        + "(?=.*\\d)"
-        + "(?=.*[@$!%*?&])"
-        + "[A-Za-z\\d@$!%*?&]{8,}$";
-    Pattern pattern = Pattern.compile(passwordRegex);
-    return pattern.matcher(password).matches();
-  }
-
-  /**
-   * Validates the user registration data.
-   *
-   * @param userRegistrationDto The user registration data to validate.
-   * @return true if the user registration data is valid, false otherwise.
-   */
-  public boolean isValidUserRegistration(UserRegistrationDto userRegistrationDto) {
-    Set<ConstraintViolation<UserRegistrationDto>> violations = validator.validate(
-        userRegistrationDto);
-
-    return violations.isEmpty();
-  }
-
-  /**
-   * Validates the user login data.
-   *
-   * @param userLoginDto The user login data to validate.
-   * @return true if the user login data is valid, false otherwise.
-   */
-  public boolean isValidUserLogin(UserLoginDto userLoginDto) {
-    Set<ConstraintViolation<UserLoginDto>> violations = validator.validate(
-        userLoginDto);
-
-    return violations.isEmpty();
-  }
 
   /**
    * Validates the user modification data.
@@ -135,71 +48,46 @@ public class UserService {
   }
 
   /**
-   * Registers a new user.
+   * Checks if the username is valid. Requirements: - Only alphanumeric characters and underscores -
+   * Length between 3 and 20 characters
    *
-   * @param userRegistrationDto The user registration data.
-   * @return A response entity with the result of the registration.
+   * @param username The username to check.
+   * @return true if the username is valid, false otherwise.
    */
-  public ResponseEntity<?> registerUser(UserRegistrationDto userRegistrationDto) {
-    // DTO validation
-    if (!isValidUserRegistration(userRegistrationDto)) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("All fields must be filled.");
-    }
-
-    // Email validation
-    if (!isValidEmail(userRegistrationDto.email())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid email address.");
-    }
-
-    // Password validation
-    if (!isValidPassword(userRegistrationDto.password())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid password.");
-    }
-
-    // Check if the username is already taken
-    if (userRepository.existsByUsername(userRegistrationDto.username())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Username already taken.");
-    }
-
-    // Check if the email is already taken
-    if (userRepository.existsByEmail(userRegistrationDto.email())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Email already taken.");
-    }
-
-    // Register the user
-    User user = userMapper.toEntity(userRegistrationDto);
-
-    // Save the user
-    userRepository.save(user);
-
-    return ResponseEntity.status(HttpStatus.CREATED).body("User registered successfully.");
+  public static boolean isValidUsername(String username) {
+    String usernameRegex = "^[A-Za-z0-9_]{3,20}$";
+    Pattern pattern = Pattern.compile(usernameRegex);
+    return pattern.matcher(username).matches();
   }
 
   /**
-   * Logs in a user.
+   * Checks if the email is valid.
    *
-   * @param userLoginDto The user login data.
-   * @return A response entity with the result of the login.
+   * @param email The email to check.
+   * @return true if the email is valid, false otherwise.
    */
-  public ResponseEntity<?> loginUser(UserLoginDto userLoginDto) {
-    // DTO validation
-    if (!isValidUserLogin(userLoginDto)) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("All fields must be filled.");
-    }
+  public static boolean isValidEmail(String email) {
+    String emailRegex = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$";
+    Pattern pattern = Pattern.compile(emailRegex);
+    return pattern.matcher(email).matches();
+  }
 
-    // Check if the user exists
-    User user = userRepository.findByUsernameOrEmail(userLoginDto.usernameOrEmail(),
-        userLoginDto.usernameOrEmail());
-    if (user == null) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User not found.");
-    }
-
-    // Check if the password is correct
-    if (!checkPassword(userLoginDto.password(), user.getPasswordHash())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Incorrect password.");
-    }
-
-    return ResponseEntity.status(HttpStatus.OK).body("User logged in successfully.");
+  /**
+   * Checks if the password is valid. Minimum requirements: - At least one uppercase letter - At
+   * least one lowercase letter - At least one digit - At least one special character - Minimum
+   * length of 8 characters
+   *
+   * @param password The password to check.
+   * @return true if the password is valid, false otherwise.
+   */
+  public static boolean isValidPassword(String password) {
+    String passwordRegex = "^(?=.*[A-Z])"
+        + "(?=.*[a-z])"
+        + "(?=.*\\d)"
+        + "(?=.*[@$!%*?&])"
+        + "[A-Za-z\\d@$!%*?&]{8,}$";
+    Pattern pattern = Pattern.compile(passwordRegex);
+    return pattern.matcher(password).matches();
   }
 
   /**
@@ -210,44 +98,66 @@ public class UserService {
    * @return A response entity with the result of the modification.
    */
   public ResponseEntity<?> updateUser(UserModifyDto userModifyDto, String username) {
+    Map<String, Object> response = new HashMap<>();
+
+    // Check if the user is trying to modify another user
+    if (!AuthenticationService.getAuthority(username)) {
+      response.put("msg", "You are not authorized to modify this user.");
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+    }
+
     // DTO validation
     if (!isValidUserModification(userModifyDto)) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("All fields must be filled.");
+      response.put("msg", "All fields must be filled.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+    }
+
+    // Username validation
+    if (!UserService.isValidUsername(userModifyDto.username())) {
+      response.put("msg", "Invalid username.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
     }
 
     // Email validation
-    if (!isValidEmail(userModifyDto.email())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid email address.");
+    if (!UserService.isValidEmail(userModifyDto.email())) {
+      response.put("msg", "Invalid email address.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
     }
 
     // Retrieve the existing user
     User existingUser = userRepository.findByUsername(username);
     if (existingUser == null) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User not found.");
+      response.put("msg", "User not found.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
     }
 
     // Check if the username is already taken
     if (!existingUser.getUsername().equals(userModifyDto.username())
         && userRepository.existsByUsername(userModifyDto.username())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Username already taken.");
+      response.put("msg", "Username already taken.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
     }
 
     // Check if the email is already taken
     if (!existingUser.getEmail().equals(userModifyDto.email())
         && userRepository.existsByEmail(userModifyDto.email())) {
-      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Email already taken.");
+      response.put("msg", "Email already taken.");
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
     }
 
     // Update the user
-    existingUser.setUsername(userModifyDto.username());
-    existingUser.setEmail(userModifyDto.email());
-    existingUser.setName(userModifyDto.name());
-    existingUser.setSurname(userModifyDto.surname());
+    userMapper.updateEntity(userModifyDto, existingUser);
 
     // Save the updated user
     userRepository.save(existingUser);
 
-    return ResponseEntity.status(HttpStatus.OK).body("User modified successfully.");
+    // Generate a JWT token
+    String jwt = jwtService.generateToken(new CustomUserDetails(existingUser));
+
+    // Return the OK status and the JWT token
+    response.put("msg", "User registered successfully.");
+    response.put("token", jwt);
+    return ResponseEntity.status(HttpStatus.OK).body(response);
   }
 
   /**
@@ -257,6 +167,12 @@ public class UserService {
    * @return A response entity with the result of the deletion.
    */
   public ResponseEntity<?> deleteUser(String username) {
+    // Check if the user is trying to delete another user
+    if (!AuthenticationService.getAuthority(username)) {
+      return ResponseEntity.status(HttpStatus.FORBIDDEN)
+          .body("You are not authorized to delete this user.");
+    }
+
     userRepository.deleteByUsername(username);
     return ResponseEntity.status(HttpStatus.OK).body("User deleted successfully.");
   }
@@ -269,5 +185,4 @@ public class UserService {
   public ResponseEntity<?> getAllUsers() {
     return ResponseEntity.status(HttpStatus.OK).body(userRepository.findAll());
   }
-
 }
